hwk3/first_service/requirements.txt
functions-framework>=3.6.0
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.18.0
